requirements.txt
ctypes-snappy==1.2
python-snappy==0.6.1
protobuf==3.20.3
requests==2.28.1
paho-mqtt==1.6.1
